fix a very rare case of seed mis-selection (and improved comments)
During the clustering process, sequence B ('father') is captured
before sequence A ('son').

Sequence B is correctly declared second in the newly formed
cluster (file 'swarm2'), with A declared as the cluster seed. This is
expected, as ties between sequences with the same number of
reads (here, 8 reads) are solved by sorting them on their name
(here, 'A' comes before 'B').

However, A appears as a cluster seed in the file 'stats2', resulting
in an KeyError downstream when building the OTU table.

The solution is to add abundance values and to sort each new cluster
before reaching the functions producing the stats2 and swarm2 output.